add user task relationship

+++ src/router/recipe.js
@@ -1,10 +1,15 @@
 const express = require('express')
 const router = new express.Router()
+const auth = require('../middleware/auth')
 const Recipe = require('../models/recipe')
 
-router.post('/recipes', async(req, res)=> {
+
+router.post('/recipes',auth, async(req, res)=> {
     try {
-        const recipe = new Recipe(req.body)
+        const recipe = new Recipe({
+            ...req.body,
+            owner: req.user._id
+        })
         await recipe.save()
         res.status(201).send(recipe)
     } catch(e) {
@@ -12,25 +17,31 @@ router.post('/recipes', async(req, res)=> {
     }
 })
 
-router.get('/recipes', async(req, res)=> {
+router.get('/recipes', auth, async(req, res)=> {
     try{
-        const recipes = await Recipe.find({})
-        res.send(recipes)
+        await req.user.populate('recipes').execPopulate()
+        if(!req.user.recipes){
+            return res.status(404).send()
+        }
+        res.send(req.user.recipes)
     } catch(e) {
         res.status(500).send()
     }
 })
 
-router.get('/recipes/:id', async(req, res)=> {
+router.get('/recipes/:id', auth, async(req, res)=> {
     try {
-        const recipe = await Recipe.findById(req.params.id)
+        const recipe = await Recipe.findOne({_id: req.params.id, owner: req.user._id})
+        if(!recipe) {
+            return res.status(404).send()
+        }
         res.send(recipe)
     } catch(e) {
         res.status(500).send()
     }
 })
 
-router.patch('/recipes/:id', async(req, res)=> {
+router.patch('/recipes/:id', auth, async(req, res)=> {
     const updates = Object.keys(req.body)
     const AllowedUpdates = ["private", "name", "ingridients"]
     const isAllowed = updates.every((update)=> AllowedUpdates.includes(update))
@@ -40,7 +51,10 @@ router.patch('/recipes/:id', async(req, res)=> {
     }
 
     try {
-        const recipe = await Recipe.findById(req.params.id)
+        const recipe = await Recipe.findOne({_id: req.params.id, owner: req.user._id})
+        if(!recipe) {
+            return res.status(404).send()
+        }
         updates.forEach((update)=> recipe[update] = req.body[update])
         await recipe.save()
         res.status(201).send(recipe)
@@ -50,9 +64,12 @@ router.patch('/recipes/:id', async(req, res)=> {
     }
 })
 
-router.delete('/recipes/:id', async(req, res)=> {
+router.delete('/recipes/:id', auth, async(req, res)=> {
     try {
-        const recipe = await Recipe.findById(req.params.id)
+        const recipe = await Recipe.findOne({_id: req.params.id, owner: req.user._id})
+        if(!recipe) {
+            return res.status(404).send()
+        }
         await recipe.delete()
         res.send(recipe)
     } catch (e) {
